continuing with the app
implemented more services
Added fluent validations

+++ Blog.Core.Application/Features/Application/UserData/UserFollow/UserFollowService.cs
@@ -2,8 +2,10 @@
 using AutoMapper;
 using Blog.Core.Application.Core;
 using Blog.Core.Application.Extensions;
+using Blog.Core.Application.Features.Application.Comments.Comments.Interfaces;
 using Blog.Core.Application.Features.Application.UserData.UserFollow.Interfaces;
 using Blog.Core.Application.Features.Application.UserData.UserFollow.Models;
+using Blog.Core.Application.Utls;
 using Blog.Core.Domain.Enums;
 
 namespace Blog.Core.Application.Features.Application.UserData.UserFollow
@@ -11,10 +13,12 @@ namespace Blog.Core.Application.Features.Application.UserData.UserFollow
     public class UserFollowService : BaseCompleteService<SaveUserFollowModel,UserModel, Domain.Entities.UserFollow>, IUserFollowService
     {
         private readonly IUserFollowRepository _userFollowRepository;
+        private readonly SessionManager _sessionManager;
 
-        public UserFollowService(IUserFollowRepository userFollowRepository, IMapper mapper) : base(userFollowRepository, mapper)
+        public UserFollowService(IUserFollowRepository userFollowRepository, IMapper mapper, SessionManager sessionManager) : base(userFollowRepository, mapper)
         {
             _userFollowRepository = userFollowRepository;
+            _sessionManager = sessionManager;
         }
         public override async Task<Result<List<UserModel>>> GetAllAsync()
         {
@@ -26,7 +30,10 @@ namespace Blog.Core.Application.Features.Application.UserData.UserFollow
         }
         public async Task<Result> DeleteAsync(string followerId, string followeId)
         {
-           bool IsDeleteASuccess = await _userFollowRepository.DeleteAsync(followerId, followeId);
+            if (!await _userFollowRepository.ExitsAsync(p => p.FollowerId == _sessionManager.GetUserFromSession().Id && p.FolloweId == followeId))
+                return ErrorTypess.NoAuthorize.Because("You do not follow this person");
+
+            bool IsDeleteASuccess = await _userFollowRepository.DeleteAsync(followerId, followeId);
 
             return !IsDeleteASuccess ? ErrorTypess.OperationFaild.Because("Error while unfollowimg the user") : Result.Success();
         }
